fix: move eslint.config.react.mjs to twenty-eslint-rules to avoid circular dependency
- Moved from packages/twenty-front/ to packages/twenty-eslint-rules/
- This avoids the circular dependency: twenty-front -> twenty-ui -> twenty-front
- Also removed root jest.config.mjs (use npx nx test instead)
