--- jest.config.mjs
@@ -1,5 +1,0 @@
-import { getJestProjects } from '@nx/jest';
-
-export default {
-  projects: getJestProjects(),
-};
