+++ packages/twenty-apps/community/linkedin-browser-extension/browser-extension/eslint.config.mjs
@@ -1,7 +1,7 @@
 import typescriptParser from '@typescript-eslint/parser';
 import path from 'path';
 import { fileURLToPath } from 'url';
-import reactConfig from '../../../../twenty-front/eslint.config.react.mjs';
+import reactConfig from '../../../../twenty-eslint-rules/eslint.config.react.mjs';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
