+++ packages/twenty-front/eslint.config.mjs
@@ -1,7 +1,7 @@
 import typescriptParser from '@typescript-eslint/parser';
 import path from 'path';
 import { fileURLToPath } from 'url';
-import reactConfig from './eslint.config.react.mjs';
+import reactConfig from '../twenty-eslint-rules/eslint.config.react.mjs';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
